% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{copyLogProposals}
\alias{copyLogProposals}
\title{Initialise the vector of Metropolis-Hastings proposals.}
\usage{
copyLogProposals(nPK, T_Prop_Theta)
}
\arguments{
\item{nPK}{number of Raman peaks in the spectral signature}

\item{T_Prop_Theta}{Vector of logarithms of the MH proposals}
}
\value{
Vector of proposals
}
\description{
This is an internal function that is only exposed on the public API for unit testing purposes.
}
