% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{methanol}
\alias{methanol}
\title{Raman spectrum of methanol (CH3OH)}
\format{A list containing 2 variables:
\describe{
  \item{wavenumbers}{a numeric Vector of 331 wavenumbers (cm^-1)}
  \item{wavenumbers}{a \code{1 * 331} Matrix of intensity values (a.u.)}
 }}
\usage{
methanol
}
\description{
Raman spectrum of methanol (CH3OH)
}
\keyword{datasets}
