% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{randomWalkVoigt}
\alias{randomWalkVoigt}
\title{Multivariate Gaussian random walk proposal.}
\usage{
randomWalkVoigt(logThetaMx, mhChol)
}
\arguments{
\item{logThetaMx}{Matrix of logarithms of the parameter values}

\item{mhChol}{Cholesky factorisation of the covariance matrix}
}
\value{
proposals for each SMC particle
}
\description{
This is an internal function that is only exposed on the public API for unit testing purposes.
}
