% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{result2}
\alias{result2}
\title{SMC particles for methanol (CH3OH)}
\format{A list containing 15 variables.}
\usage{
result2
}
\description{
Posterior distribution for pseudo-Voigt parameters, obtained by running
`fitVoigtPeaksSMC` on a Raman spectrum of methanol with 4 peaks.
For details, refer to the vignette.
}
\keyword{datasets}
