% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TAMRA}
\alias{TAMRA}
\title{Surface-enhanced Raman spectram of tetramethylrhodamine+DNA (T20)}
\format{A ``hyperSpec`` spc object}
\source{
\url{https://pure.strath.ac.uk/portal/files/43595106/Figure_2.zip}
}
\usage{
TAMRA
}
\description{
Surface-enhanced Raman spectram of tetramethylrhodamine+DNA (T20)
}
\keyword{datasets}
