% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{weightedGaussian}
\alias{weightedGaussian}
\title{Compute the spectral signature using Gaussian peaks.}
\usage{
weightedGaussian(location, scale, amplitude, wavelengths)
}
\arguments{
\item{location}{Vector of location parameters of the peaks (mean).}

\item{scale}{Vector of scale parameters of the peaks (standard deviation).}

\item{amplitude}{Vector of amplitudes of the peaks.}

\item{wavelengths}{Vector of wavenumbers at which to compute the function.}
}
\value{
The value of the Gaussian function at the given wavelengths.
}
\description{
Calculates the value of the squared exponential radial basis function at the given wavelengths,
given the parameters of the peaks. This function is thread-safe.
}
