% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{weightedLorentzian}
\alias{weightedLorentzian}
\title{Compute the spectral signature using Lorentzian peaks.}
\usage{
weightedLorentzian(location, scale, amplitude, wavelengths)
}
\arguments{
\item{location}{Vector of location parameters of the peaks.}

\item{scale}{Vector of scale parameters of the peaks.}

\item{amplitude}{Vector of amplitudes of the peaks.}

\item{wavelengths}{Vector of wavenumbers at which to compute the function.}
}
\value{
The value of the Lorentian function at the given wavelengths.
}
\description{
Calculates the value of the Lorentzian function at the given wavelengths,
given the parameters of the peaks. This function is thread-safe.
}
