% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{weightedMean}
\alias{weightedMean}
\title{Compute the weighted arithmetic means of the particles.}
\usage{
weightedMean(particles, log_weights)
}
\arguments{
\item{particles}{\code{npeaks * npart} Matrix of parameter values for each particle.}

\item{log_weights}{logarithms of the importance weights of each particle.}
}
\value{
A vector of means, one for each row.
}
\description{
This SMC estimate of the means can be used to centre independent Metropolis-Hastings proposals.
}
\seealso{
\code{\link[stats:weighted.mean]{weighted.mean}}
}
