% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{weightedVariance}
\alias{weightedVariance}
\title{Compute the weighted variance of the particles.}
\usage{
weightedVariance(particles, log_weights, mean)
}
\arguments{
\item{particles}{\code{npeaks * npart} Matrix of parameter values for each particle.}

\item{log_weights}{logarithms of the importance weights of each particle.}

\item{mean}{Vector of weighted means of each particle.}
}
\value{
A vector of variances, one for each row.
}
\description{
This SMC estimate of the variance can be used to scale the bandwidth of adaptive,
Gaussian random walk Metropolis-Hastings proposals.
}
\seealso{
\code{\link[Hmisc:wtd.stats]{wtd.var}}
}
