% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lsTamra}
\alias{lsTamra}
\title{Surface-enhanced Raman spectram of tetramethylrhodamine+DNA (T20)}
\format{
A list containing 2 variables:
\describe{
  \item{wavenumbers}{a numeric Vector of 2401 wavenumbers (cm^-1)}
  \item{wavenumbers}{a \code{1 * 2401} Matrix of intensity values (a.u.)}
 }
}
\source{
Gracie et al. (Anal. Chem., 2016), Figure 2, spectrum T20_1 REP1 Well A1,
  restricted to 600--1800 cm^-1. Dataset \doi{10.15129/2c186689-2334-4746-955e-f9579ca15909}
}
\usage{
lsTamra
}
\description{
Surface-enhanced Raman spectram of tetramethylrhodamine+DNA (T20)
}
\keyword{datasets}
